Pack a tuple with the success status when packing the API result object
This ensures us that the client knows when something crashed, by checking
the status index. Now we are sending a two-pair tuple object, containing
the status as an integer, where 0 designates success and 1 designates failure
and the second object being either the actual result or the error message
in case of an error.