Add a thread for collecting the stale connections
We can have stale connections whenever a client does not close the connection
on its end. The connection might not be actually stale, but it can be a long lived
connection, which the client keeps opened with the purpose of reusing it later on.

This possibility, though, creates a huge problem whenever we are quitting foobar,
since the component will get stuck on waiting the threads it started to finished,
the threads being themselves stucked, usually waiting through WaitForSingleObject
for a new action to occur from the client's side. In order to safely close foobar2000
and its corresponding resources, we are tracking those long lived connections
and explicitly close them, making the threads to eventually stop.
This action, though, creates a problem, since each of those threads will try to
close their connection when they can't retrieve anything from it, leading to a runtime
error due to closing the same handle twice. In order to alleviate this, we use a single
point of processing those connections, which is a thread responssible for retrieving
the connections from a queue and closing them as soon as it gets its hand on them.